requirements.txt
docker
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
pytest
sqlparse
djangorestframework